Cloud Custodian Mailer Cheatsheet
Custodian mailer is to install the mailer, and run a policy that triggers an email to your inbox.

Steps: [Install] the mailer on your laptop.
In your text editor, create a `mailer.yml` file to hold your mailer config.
In the AWS console, create a new standard SQS queue (quick create is fine).
Copy the queue URL to `queue_url` in `mailer.yml`.
In AWS, locate or create a role that has read access to the queue. Grab the role ARN and set it as `role` in `mailer.yml`.
Make sure your email address is verified in SES, and set it as `from_address` in `mailer.yml`. By default SES is in sandbox mode where you
must.
Also set `region` if you are in a region other than `us-east-1`.
Create a Custodian policy to populate your mailer queue. Create a `test-policy.yml` file with the content (update `to` and `queue` to match your environment).
Then Run custodian Mailer command against your environment.